requirements.txt
TinyDB
python-telegram-bot==13.15
requests
json